requirements.txt
discord.py==0.16.12
requests>=2.20.0
requests-oauthlib==1.0.0
praw==6.0.0
prawcore==1.0.0
pandas>=0.23.4